feat(li): function based attribution is added, dimension drill down attribution is throughly revised
